refactor: shared로 이동
